Refactor device_refresh script to use new DeviceRefresh class
The existing standalone device_refresh script has been refactored to use a newly created DeviceRefresh class. This change removes redundant code and structures from the script, keeping it more clean and manageable. The DeviceRefresh class encapsulates the logic which was spread across multiple parts of the script in a cohesive and reusable manner. The script now also uses an instance of the custom PanOsUpgradeLogger.