Refactored upgrade process for HA and standalone devices
The code for the PAN-OS upgrade has been refactored to make it more robust and organized. Added exception handlers for general and specific errors during the upgrade process for devices in an High Availability pair or standalone. This refactoring also led to a significant reduction in the lines of code. Log messages and error handling have both been improved, so troubleshooting the upgrade process will be easier in the future.