Implement PAN-OS version syncing from devices
The upgrade-list component has been updated to actually sync PAN-OS versions from a selected device. Modification includes calling the syncPanosVersions() method from the UpgradeService when the user opts to sync versions from the device. Observable syncVersions$ has been added to monitor the progress of the sync operation. The sync operation is now tied to both selected devices and a selected profile, and not just the devices.