Refactor upgrade service to add syncPanosVersions method
The upgrade service is simplified and refactored. Obsolete methods related to upgrade history, status, and canceling have been removed. A new method 'syncPanosVersions' is introduced for synchronizing PANOS versions. This also modifies the existing 'getPANOSVersions' method to 'getPanosVersions'. This newly introduced method empowers the service with synchronous operations to ensure up-to-date data.