Update service configurations in Helm chart
This revision involves a Helm Chart update that modifies the Kubernetes service configurations for the backend, frontend, database, and cache components of the pan-os-upgrade-web application. Changes primarily include the addition of labels and tweaks to the selector labels for better identity and organisation. It also allows for dynamic port assignments for the node service type in frontend.