Update values in pan-os-upgrade-web/values.yaml
This commit adjusts a few configuration settings within the values file of pan-os-upgrade-web. It modifies the CELERY_BROKER_URL, CELERY_RESULT_BACKEND, and DATABASE_URL values to use environment variables rather than hardcoded values. The 'worker' block is also moved to the end of the file, and additional service information for the frontend is defined.