Add Documentation link to sidenav
This commit introduces a new link to the side navigation bar for the 'Documentation' section. The update adds a text and an icon link for Documentation and separates it from other links with a divider. Indentation of the 'Home' link is additionally corrected in order to maintain the code's readability.