Refactor upgrade-diagram component to use UpgradeStepService
The upgrade-diagram component has been modified to use UpgradeStepService for fetching the SVG path of the current step. As part of this change, the component no longer implements OnChanges. The SVG path is now obtained asynchronously, with image loading and visibility controlled through async pipe in the template. This change simplifies code and improves component efficiency.