Add UpgradeStepService to manage upgrade steps
This newly added service, UpgradeStepService, is designed to manage the different steps involved in an upgrade process. It provides mapping for each step with their corresponding SVG graphics and allows the current step to be updated based on log messages. This allows for real-time tracking and visualization of upgrade steps.