Simplify update of 'Errored' status in upgrade process
This commit streamlines the process of updating the 'Errored' status during the upgrade process. Instead of using a specific targeted device hostname to update the current step of the upgrade job, the upgrade job will simply use a general 'Errored' status. This makes the code cleaner and easier to maintain.