Add device details to job detail page
The device details object has been added in job-details.ts to be accessed by different sections of the job detail page. Job details' value was previously accessing from the job object directly, but now it has been shifted to the device object in the job object which provides detailed information about the device.