Update error handling in upgrade_device script
The upgrade_device script in the panosupgradeweb app has been updated to better handle errors. This is achieved by including more detailed information when updating the current step of the upgrade job in the case of an error. Now, the device name and step name are specified, providing better context for diagnosing and addressing issues.