Refactor job creation in tasks.py
The way job objects are created in tasks.py has been refactored. Now, a dictionary named 'job_data' is used to group all parameters related to the job. Additionally, the process includes a check for a peer device, and if it exists, the corresponding data is also added to the 'job_data' dictionary.