Add device name to the update step in upgrade process
The update_current_step method has been modified to now include the name of the device that the upgrade is relating to. This change gives more context to each step being performed in the upgrade process, and makes it easier to identify which device a certain action or error message is associated with.