Update job service to use JobStatus interface
The changes involve modifying all Job type declarations in the job service to JobStatus. This shift will ensure that all job retrieval and creation operations now return or make use of the JobStatus interface in place of previously used Job interface. This makes our interpretation of Jobs more consistent with the current architecture of the system.