Update naming and data source of diagram component
The upgrade diagram component has been updated to change the source of device information from 'deviceDetails' to 'jobStatusDetails'. This is reflected in the HTML and TypeScript files. Now, the device specific details and status are derived from the job status, rather than directly from the device details. Changes are made to enhance the flow of information and improve code consistency.