Add target_current_status to job data
In this update, a new attribute "target_current_status" was introduced in the job data dict. This attribute provides the status of the target device during the upgrade process. Throughout the code, current status is now captured at various stages i.e., active, errored, and completed. This helps in better tracking the status of upgrade jobs.