Refactor upgrade diagram component styling and structure
The main changes revolved around modifying the layout and styling of the upgrade diagram component. SCSS rules were updated to shift from flex to column layout, optimize width and height specifications, and modify display for firewall details. Additionally, the HTML template and typescript file saw changes to accommodate the styling revisions, notably the incorporation of MatChipsModule and the creation of a new method 'hasPeerDevice.'