Update padding and width in job-details styling
The styles for job-details have been updated to improve visual balance. The padding in the job-details-container was increased from 20px to 40px for more white space, and the width in job-details has been commented out to check for potential layout issues. Moreover, redundant job-details styling found in _job-details-theme.scss was removed to maintain clean and concise code.